Rework static file serving; add support for serving media
Changes:

- Remove WhiteNoise and rely on an external server for serving static
  files in production, since such an external server would be needed
  anyway to serve uploaded media.

- Support serving uploaded media through the following strategies:
  - When DEBUG=true, serve media from Hummus itself by default. This can
    be disabled by setting SERVE_MEDIA=false.
  - When DEBUG=false (or DEBUG=true and SERVE_MEDIA=false), use
    X-Accel-Redirect to instruct the proxy to serve uploaded media
    files. (This allows Hummus to check that the user is authenticated
    before the proxy sends the files.) The base URL to use for this is
    configured by MEDIA_X_ACCEL_REDIRECT_URL.
  - When neither of the above is configured, nothing is done.

- Make the docker-compose development environment use an nginx reverse
  proxy so that we can test X-Accel-Redirect.

- Move the Dockerfile and any development environment configuration
  files (currently, just a configuration file for nginx, introduced in
  this commit) into its own directory deploy/dev.

The provided Dockerfile is no longer suitable for building a production
image. A new one will have to be provided---or rather, I intend to add
deploy/compose and deploy/kubernetes directories containing resources
for deploying Hummus using docker-compose and Kubernetes, respectively.
In the meantime, README.md is updated to note the fact that there's no
good way to deploy into production right now.

README.md will also have to be updated with better documentation of the
new configuration options once the production deployment options are
available. (What I have in this commit is sort of a stop-gap.)