Use ManifestStaticFilesStorage
Now static files can be cached indefinitely in production.

The application container image must now include the Django
collectstatic management command as a build step, since this creates the
staticfiles.json file that Django needs to compute static file URLs. The
existing development Dockerfile/docker-compose setup would still be fine
(since it is used with DEBUG=true and ManifestStaticFilesStorage acts
like the default file storage in this case), but I updated it so that it
will be with the forthcoming production Dockerfiles.